Improves modularity
