+++ index.html
@@ -228,5 +228,7 @@
   </template>
 
 <script type="module" src="js/main.js"></script>
+<script src="js/function.js"></script>
+
 </body>
 </html>
