+++ js/main.js
@@ -1,2 +1,1 @@
-import './functions.js';
 import './get-photos.js';
